// Nijo.GuiClient/package_ui-components/src/layout/DynamicForm/types.ts
export type DynamicFormProps<TRoot extends ReactHookForm.FieldValues = ReactHookForm.FieldValues> = FormLayoutProps & {
  /** データ構造の定義 */
  root: MemberOwner<TRoot>
  /** フォームのデフォルト値 */
  defaultValues?: ReactHookForm.DefaultValues<TRoot>
  /** フォームを読み取り専用にするかどうか */
  isReadOnly?: boolean
}

/** DynamicFormのref */
export type DynamicFormRef<TRoot extends ReactHookForm.FieldValues = ReactHookForm.FieldValues> = {
  /** react-hook-form の useForm の戻り値。最新の値の取得などに使用する。 */
  useFormReturn: ReactHookForm.UseFormReturn<TRoot>
  /** フォーム定義をコンソール出力する（デバッグ用） */
  consoleLog: () => void
}

// ----------------------------------------------

/** ルートオブジェクト、ネストされたオブジェクト、配列、のいずれか */
export type MemberOwner<TOwner extends ReactHookForm.FieldValues = ReactHookForm.FieldValues> = {
  /** メンバー */
  members: Member<TOwner>[]
}

// ----------------------------------------------
//#region メンバー

/** DynamicFormのメンバー */
export type Member<TOwner extends ReactHookForm.FieldValues = ReactHookForm.FieldValues> =
  | SectionMember<TOwner>
  | ArrayMember<TOwner>
  | ValueMember<TOwner>
// ...
export type ValueMember<TOwner extends ReactHookForm.FieldValues = ReactHookForm.FieldValues> = {
  type?: never
  /** このフィールドのオーナーからこの値メンバーまでのパス */
  physicalName?: ReactHookForm.Path<TOwner>
  /** フォームのラベル部分 */
  label?: string | CustomValueRenderer
  /** 値部分のレンダリングコンポーネント。未指定の場合は何も表示されなくなる。 */
  contents?: CustomValueRenderer
  /** グリッドの列定義。未指定の場合はグリッドに表示されなくなる。 */
  getGridColumnDef?: GetGridColumnDefFunction
  /** このメンバーを横幅いっぱいにするかどうか。 */
  fullWidth?: boolean
}

//#endregion メンバー

// ----------------------------------------------
//#region フォームのカスタマイザー

// 各種レンダリング関数
export type CustomRenderer<T extends FormRendererProps> = (props: T) => React.ReactNode
export type CustomSectionRenderer = CustomRenderer<SectionFormRendererProps>
export type CustomArrayRenderer = CustomRenderer<ArrayFormRendererProps>
export type CustomArrayItemRenderer = CustomRenderer<ArrayFormItemRendererProps>
export type CustomValueRenderer = CustomRenderer<ValueMemberFormRendererProps>
// ...
export type ValueMemberFormRendererProps = FormRendererProps & {
  /** ルートオブジェクトからこのメンバーまでのパス */
  name: string
}

//#endregion フォームのカスタマイザー
// ----------------------------------------------
//#region グリッド

/** 値メンバーのグリッドの列定義を取得するための関数 */
export type GetGridColumnDefFunction = (props: GetGridColumnDefFunctionProps) => GridColumnDef

/** グリッドの列定義を取得するための関数の引数 */
export type GetGridColumnDefFunctionProps = {
  /** ルートオブジェクトから配列までのパス */
  arrayPath: string
  /** 配列の行オブジェクトからこのメンバーまでのパス */
  pathFromRow: string
  /** メンバーのメタデータ */
  member: ValueMember
  /** メンバーを保持するオブジェクトのメタデータ */
  owner: MemberOwner
  /** グリッドの列定義を作成するためのヘルパー関数 */
  cellType: ColumnDefFactories<ReactHookForm.FieldValues>
  /** フォームが読み取り専用かどうか */
  isReadOnly: boolean
}

/** グリッドの列定義 */
export type GridColumnDef = EditableGridColumnDef<ReactHookForm.FieldValues>